[Frontend] Refatoracao de estrutura de arquivos feito para categorias e produtos.
