+++ back_end/actions.py
@@ -3,26 +3,15 @@ import sys
 sys.path.append('.')
 
 from back_end.log import generate_log, read_logs
-import back_end.db as db
+import back_end.controller.ctrl_category as db
+import back_end.controller.ctrl_product as db_prod
+
 
 
 def create_marketplace(name: str, description: str):
     marketplace = f'{name};{description}'
     db.add_marketplace(marketplace)
     generate_log(f'Cadastro do marketplace "{name}" ao database.')
-
-
-def create_product(name: str, description: str, price: float):
-    product = f'{name};{description};{price}'
-    db.add_product(product)
-    generate_log(f'Cadastro do produto "{name}" ao database.')
-
-
-def create_category(name: str, description: str):
-    category = f'{name};{description}'
-    db.add_category(category)
-    generate_log(f'Cadastro da categoria "{name}" ao database.')
-
 
 def create_seller(name: str, phone: str, email: str):
     seller = f'{name};{phone};{email}'
@@ -36,24 +25,12 @@ def list_marketplaces():
     return marketplaces
 
 
-def list_products():
-    products = db.read_products()
-    generate_log('Listado todos os produtos.')
-    return products
-
-
 def list_sellers():
     sellers = db.read_seller()
     generate_log('Listado todos os sellers.')
     return sellers
 
-
-def list_categories():
-    categories = db.read_categories()
-    generate_log('Listado todas as categorias.')
-    return categories
-
-
+#pode ser tirado do codigo
 def list_logs():
     logs = read_logs()
     return logs
