+++ back_end/db.py
@@ -4,28 +4,7 @@ sys.path.append('.')
 from back_end.log import generate_log
 
 marketplaces_txt = 'data/marketplaces.txt'
-product_db = 'data/product.txt'
-categories_db = 'data/categories.txt'
 sellers_db = 'data/sellers.txt'
-
-
-def add_product(product: str) -> None:
-    with open(product_db, 'a', encoding='utf-8') as products_file:
-        products_file.write(f'{product}\n')
-
-
-def read_products() -> list:
-    with open(product_db, 'r', encoding='utf-8') as products_file:
-        products = []
-        for product in products_file:
-            name, description, price = product.strip().split(';')
-            products.append({'name': name, 'description': description, 'price': float(price)})
-        return products
-
-
-def verify_db():
-    with open(product_db, 'r', encoding='utf-8') as products_file:
-        return products_file
 
 
 def add_marketplace(marketplace: str) -> None:
@@ -42,22 +21,6 @@ def read_marketplace() -> list:
 
     return marketplaces
 
-
-def add_category(category: str) -> None:
-    with open(categories_db, 'a', encoding='utf-8') as categories_file:
-        categories_file.write(f'{category}\n')
-
-
-def read_categories() -> list:
-    with open(categories_db, 'r', encoding='utf-8') as categories_file:
-        categories = {}
-        for category in categories_file:
-            name, description = category.strip().split(';')
-            categories[name] = description
-
-    return categories
-
-
 def add_seller(seller: str) -> None:
     with open(sellers_db, 'a', encoding='utf-8') as sellers_file:
         sellers_file.write(f'{seller}\n')
@@ -73,11 +36,4 @@ def read_seller() -> list:
     return sellers
 
 
-def read_categories() -> list:
-    with open(categories_db, 'r', encoding='utf-8') as categories_file:
-        categories = {}
-        for category in categories_file:
-            name, description = category.strip().split(';')
-            categories[name] = description
 
-    return categories
