+++ data/product.txt
@@ -5,3 +5,4 @@ boneca;brinquedo;1
 boneca;brinquedo;1
 boneca;brinquedo;3
 boneca;brinquedo;5
+teste produto 11 01 2021;produto bonito;34
