+++ front_end/web.py
@@ -2,7 +2,9 @@ import sys
 
 sys.path.append('.')
 from flask import Flask, render_template, request
-import back_end.actions as actions
+import back_end.controller.ctrl_category as ct_category
+import back_end.controller.ctrl_product as ct_product
+import back_end.dao.dao_log as dao_ac_log
 
 app = Flask(__name__)
 titulo_head = 'Lojinha'
@@ -26,7 +28,7 @@ def cadastro_Produto():
     if product_name is not None:
         product_description = request.args.get('description')
         product_price = request.args.get('price')
-        actions.create_product(product_name, product_description, product_price)
+        ct_product.create_product(product_name, product_description, product_price)
         menssagem = f'{product_name} cadastrado com sucesso'
     return render_template('create_product.html', menssagem=menssagem, titulo='Cadastro de Produtos', titulo_head=titulo_head)
 
@@ -38,7 +40,7 @@ def cadastro_Categoria():
 
     if category_name is not None:
         category_description = request.args.get('description')
-        actions.create_category(category_name, category_description)
+        ct_category.create_category(category_name, category_description)
         mensagem = f'{category_name} cadastrado com sucesso'
     return render_template('create_category.html', menssagem=mensagem, titulo='Cadastro de Categorias', titulo_head=titulo_head)
 
@@ -70,20 +72,18 @@ def lista_sellers():
 
 @app.route('/listar_produtos')
 def lista_produtos():
-    products = actions.list_products()
+    products = ct_product.list_products()
     return render_template('list_products.html', products=products, titulo="Produtos", titulo_head=titulo_head)
-
 
 @app.route('/listar_categorias')
 def lista_categorias():
-    categories = actions.list_categories()
+    categories = ct_category.list_categories()
     return render_template('list_categories.html', categories=categories, titulo="Categorias", titulo_head=titulo_head)
-
 
 
 @app.route('/listar_logs')
 def lista_logs():
-    logs = actions.list_logs()
+    logs = dao_ac_log.read_logs()
     return render_template('list_logs.html', logs=logs, titulo="Histórico", titulo_head=titulo_head)
 
 
@@ -92,4 +92,4 @@ def home():
     return render_template('home.html', titulo_head=titulo_head)
 
 
-app.run(debug=True)
+app.run()
